NewsFeed

+++ Source/HRHiringTool/HRHiringTool/Classes/OpeningNote.cs
@@ -8,8 +8,8 @@ namespace HRHiringTool.Classes
     public class OpeningNote
     {
         public string Note { get; set; }
-        public int IdJobOpening { get; set; }
-        public int IdUser { get; set; }
+        public long IdJobOpening { get; set; }
+        public long IdUser { get; set; }
         public string UserName { get; set; }
         public DateTime Timestamp { get; set; }
     }
